requirements.txt
joblib==1.4.2
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.3
shap==0.47.2
scikit-learn==1.5.1
streamlit==1.39.0
transformers==4.52.4
xgboost==2.1.4
yfinance==0.2.61
newsapi-python  # No version available programmatically; install latest
